Add a port for clustershell
Clustershell is a program that lets you run commands, and transfer files
to/from, many servers at the same time.  It's meant for HPC environments, but
works anywhere you have multiple machines (and can list them).

The code lives in Github, so we pull from there directly, rather than going
through pip or another repository.

ClusterShell is hard-coded to look in /etc/clustershell for system-wide
configuration, so we patch that to look in ${prefix}.