New port: libguestfs-appliance
This is a custom-built virtual appliance that many libguestfs
tools use to examine and manipulate disk images.

When installed on a Linux system with all required dependencies,
libguestfs uses supermin(1) to build very small virtual appliances
to do the job covered by this pre-built appliance.

When supermin's dependencies are not going to be available, the
tool libguestfs-make-fixed-appliance(1) can be used to create a
larger virtual appliance that is able to handle all of the tasks
of the smaller supermin(1) appliances.

Unfortunately, libguestfs-make-fixed-appliance can only be run on
Linux.  For those situations, Red Hat distributes already-built
virtual appliances that can be run using QEMU.  The appliance
contains a minimal Fedora install, along with the guestfsd(8)
daemon.

libguestfs does not require the appliance at build time, and some
libguestfs commands can be run without it, but most operations
require either supermin, or the appliance.